requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.22.0
xformers>=0.0.20
numpy>=1.24.0
pillow>=10.0.0
matplotlib>=3.7.0
tqdm>=4.66.0
pyyaml>=6.0
monai>=1.2.0
scikit-image>=0.20.0
opencv-python>=4.8.0
clean-fid>=0.1.35
wandb>=0.15.0
torchmetrics>=1.0.0 